+++ R/segregation.R
@@ -13,9 +13,10 @@ NULL
 
 globalVariables(c(
     "V1", "V2", "cond1", "cond2", "entropy_cond", "entropy_cond1", "entropy_cond2", "entropyw",
-    "est", "freq", "freq1", "freq2", "freq_orig1", "freq_orig2", "ls_diff1", "ls_diff2", "ls_unit",
+    "est", "freq", "freq1", "freq2", "freq_orig1", "freq_orig2",
+    "ls_diff_mean", "ls_diff1", "ls_diff2", "ls_unit", "bias", "mean_boot",
     "n", "n_group", "n_group_target", "n_source", "n_target", "n_unit", "n_unit_target",
-    "n_within_group", "p", "p1", "p2", "p_group", "p_group_g_unit", "p_group_g_unit1",
+    "n_within_group", "p", "p_exp", "p1", "p2", "p_group", "p_group_g_unit", "p_group_g_unit1",
     "p_group_g_unit2", "p_group_s", "p_group_t", "p_unit", "p_unit1", "p_unit2", "p_unit_s",
     "p_unit_t", "p_within", "sumcond1", "sumcond2", "total", "unit1", "unit2"))
 
@@ -79,7 +80,9 @@ logf <- function(v, base) {
     if (missing(base)) {
         stop("argument base required")
     }
-    ifelse(v > 0 & is.finite(v), log(v, base = base), 0)
+    logged <- log(v, base = base)
+    logged[!is.finite(logged)] <- 0
+    logged
 }
 
 #' Calculates the entropy of a distribution
@@ -135,11 +138,14 @@ prepare_data <- function(data, group, unit, weight, within = NULL) {
     }
     vars <- c(group, unit)
 
+    # create a copy
+    data <- as.data.table(data)
+
     # use provided weight or weight of 1
     if (!is.null(weight)) {
-        data[, "freq"] <- as.double(data[[weight]])
+        data[, freq := as.double(get(weight))]
     } else {
-        data[, "freq"] <- 1
+        data[, freq := 1]
     }
 
     if (!is.null(within)) {
@@ -147,7 +153,6 @@ prepare_data <- function(data, group, unit, weight, within = NULL) {
     }
 
     # collapse on vars, and select only positive weights
-    data.table::setDT(data)
     data <- data[freq > 0, list(freq = sum(freq)), by = vars]
     setattr(data, "vars", vars)
     setkey(data, NULL)