+++ NEWS.md
@@ -1,3 +1,10 @@
+# segregation 0.4.0
+
+* faster bootstrap
+* return bootstrap estimates as attr
+* add mutual_expected
+* apply bias-correction via bootstrap by default when se=TRUE
+
 # segregation 0.3.0
 
 * always return data.table